requirements.txt
asgiref==3.8.1
astroid==3.2.3
dill==0.3.8
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
Faker==26.0.0
isort==5.13.2
Markdown==3.6
mccabe==0.7.0
platformdirs==4.2.2
pylint==3.2.5
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.0
validate-docbr==1.10.0
